Move `RatesViewModel` to `provider.ui.viewmodel` package.
The `RatesViewModel` class has been moved from `com.currency.rateman.ui.viewmodels` to `com.currency.rateman.provider.ui.viewmodel`.
Imports in `ProviderListScreen.kt`, `TargetCurrencyScreen.kt`, and `KoinModules.kt` have been updated to reflect this change.
