+++ app/src/main/java/com/currency/rateman/di/KoinModules.kt
@@ -8,7 +8,7 @@ import com.currency.rateman.data.repository.SettingsRepository
 import com.currency.rateman.data.repository.SettingsRepositoryImpl
 import com.currency.rateman.ui.viewmodels.CurrencyViewModel
 import com.currency.rateman.provider.ui.viewmodel.ProviderDetailViewModel
-import com.currency.rateman.ui.viewmodels.RatesViewModel
+import com.currency.rateman.provider.ui.viewmodel.RatesViewModel
 import com.currency.rateman.ui.viewmodels.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
